feat: provide all event data to agent in JSON format feat: env var to enable/disable near worker fix: code review

+++ hub/tasks/neardata.py
@@ -1,4 +1,5 @@
 import json
+import logging
 import os
 import re
 from contextlib import asynccontextmanager
@@ -8,10 +9,14 @@ from typing import Any
 import httpx
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from apscheduler.triggers.interval import IntervalTrigger
+from dotenv import load_dotenv
 from nearai.shared.client_config import DEFAULT_MODEL
 
 from hub.api.v1.auth import AuthToken
 from hub.api.v1.thread_routes import RunCreateParamsBase, ThreadModel, _create_thread, create_run
+
+load_dotenv()
+logger = logging.getLogger(__name__)
 
 RESET_BLOCK_ID_ON_START = True
 
@@ -44,7 +49,7 @@ async def get_latest_block_id():
     if data:
         latest_block_id = data.get("sync_block_height", None)
         if latest_block_id:
-            print(f"Starting with the latest block ID: {latest_block_id}")
+            logger.info(f"Starting with the latest block ID: {latest_block_id}")
             return latest_block_id
 
     return 0
@@ -60,14 +65,14 @@ async def get_last_block_id():
 
 # Function to save the last blockId to a file
 def save_last_block_id(block_id):
-    print("Saving last block id", block_id)
+    logger.info(f"Saving last block {block_id}")
     with open(BLOCK_ID_FILE, "w") as f:
         f.write(str(block_id))
 
 
 # Asynchronous function to load a block by its ID
 async def load_block(block_id, auth_token):
-    print(f"Loading block {block_id}")
+    logger.info(f"Loading block {block_id}")
     data = await async_fetch_json(f"{BASE_URL}{block_id}")
     if data and data.get("shards"):
         # If data and shards are found, process them
@@ -99,23 +104,14 @@ async def process_log(log, receipt_execution_outcome, auth_token):
             for data in event["data"]:
                 # Check the event type and call the appropriate save_event function
 
-                max_iterations = data.get("max_iterations", 1)
-                if max_iterations:
-                    if isinstance(max_iterations, (int, float)) and max_iterations > 0:
-                        max_iterations = int(max_iterations)
-                    else:
-                        max_iterations = 1
-                else:
-                    max_iterations = 1
-
                 if event.get("event") == "run_agent":
                     await run_agent(
                         data.get("agent", ""),
-                        data.get("message", ""),
-                        data.get("signer_id", ""),
                         {
-                            "max_iterations": max_iterations,
-                            "env_vars": data.get("env_vars", "{}"),
+                            "event": event.get("event"),
+                            "message": data.get("message", ""),
+                            "signer_id": data.get("signer_id", ""),
+                            "env_vars": data.get("env_vars", None),
                             "referral_id": data.get("referral_id", ""),
                             "amount": data.get("amount", ""),
                             "receipt_id": receipt_execution_outcome["execution_outcome"]["id"],
@@ -135,11 +131,11 @@ def load_auth_token():
     return app_config.auth
 
 
-async def run_agent(agent, message, signer_id, data, auth_token: AuthToken):
-    if not (agent and message):
-        return print("Missing data")
+async def run_agent(agent, content, auth_token: AuthToken):
+    if not agent:
+        return logger.error("Missing data")
 
-    # TODO find out how to ret fid of 401 error here
+    # TODO find out how to get rid of 401 error here
     # auth_token.on_behalf_of = signer_id
 
     thread_model = ThreadModel(
@@ -160,7 +156,7 @@ async def run_agent(agent, message, signer_id, data, auth_token: AuthToken):
         metadata=None,
         include=[],
         additional_instructions=None,
-        additional_messages=[{"content": message, "role": "user"}],
+        additional_messages=[{"content": json.dumps(content), "role": "user"}],
         max_completion_tokens=None,
         max_prompt_tokens=None,
         parallel_tool_calls=None,
@@ -173,7 +169,6 @@ async def run_agent(agent, message, signer_id, data, auth_token: AuthToken):
         schedule_at=None,
         delegate_execution=False,
         parent_run_id=None,
-        # max_iterations=data.get("max_iterations", 1),
     )
 
     create_run(thread_id=thread.id, run=run_params, auth=auth_token, scheduler=scheduler)
@@ -188,11 +183,11 @@ async def process_blocks(auth_token):
     new_block_id = await load_block(block_id + 1, auth_token)
     if new_block_id:
         # If block is found, save the new blockId
-        print(f"New block found: {new_block_id}")
+        logger.info(f"New block found: {new_block_id}")
         save_last_block_id(new_block_id)
         return True
     else:
-        print(f"Block not found: {block_id}")
+        logger.warning(f"Block not found: {block_id}")
         return False
 
 
@@ -206,19 +201,22 @@ async def periodic_task(auth_token: AuthToken):
 
 @asynccontextmanager
 async def lifespan(app):
-    # remove last known block_id to reset the state on start
-    if RESET_BLOCK_ID_ON_START:
-        if os.path.exists(BLOCK_ID_FILE):
-            os.remove(BLOCK_ID_FILE)
-            print(f"File {BLOCK_ID_FILE} has been deleted.")
+    if os.getenv("READ_NEAR_EVENTS", "false").lower() == "true":
+        # remove last known block_id to reset the state on start
+        if RESET_BLOCK_ID_ON_START:
+            if os.path.exists(BLOCK_ID_FILE):
+                os.remove(BLOCK_ID_FILE)
+                logger.info(f"File {BLOCK_ID_FILE} has been deleted.")
 
-    auth_token = load_auth_token()
+        auth_token = load_auth_token()
 
-    job = partial(periodic_task, auth_token)
+        job = partial(periodic_task, auth_token)
 
-    scheduler.add_job(job, IntervalTrigger(seconds=1))
-    scheduler.start()
+        scheduler.add_job(job, IntervalTrigger(seconds=1))
+        scheduler.start()
 
-    yield
+        yield
 
-    scheduler.shutdown()
+        scheduler.shutdown()
+    else:
+        yield
